Update 2DarrayRecursiveSumPath(NOT WORKING).html
I fixed the problem. 

Main issues were:
(1) changed: if(pathObj.j < a[pathObj.i].length) to if(pathObj.j < a[pathObj.i].length - 1)
(2) changed: if(pathObj.i < a.length ) to if(pathObj.i < a.length - 1)

This makes no sense to me. The previously changed statements were acting out as a <= statement. Therefore, I'll be so bold as to say that javascript is incorrect.

Check my C++ solution in this same repository. I did not have to change this statement and it works as it should.